# src/app_server/straight-as-api/apidoc.json
{
  "openapi" : "3.0.0",
  "servers" : [ {
    "description" : "SwaggerHub API Auto Mocking",
    "url" : "https://virtserver.swaggerhub.com/Architecton/StraightAs/1.0.0"
  } ],
  "info" : {
    "description" : "Documentation for the StraightAs API\n",
    "version" : "1.0.0",
    "title" : "StraightAs"
  },
...
  "paths" : {
    "/users" : {
      "get" : {
        "tags" : [ "Users" ],
        "summary" : "Get all users in the database",
        "operationId" : "userGetAll",
        "responses" : {
          "200" : {
            "description" : "Users were successfully retrieved",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/User"
                }
              }
            }
          },
          "404" : {
            "description" : "Cannot find users"
          },
          "500" : {
            "description" : "Error while getting users"
          }
        }
      },
      "post" : {
        "tags" : [ "Authentication" ],
        "summary" : "Add a new user to the database",
        "operationId" : "userAddNew",
        "responses" : {
          "201" : {
            "description" : "User successfully created",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/User"
                }
              }
            }
          },
          "400" : {
            "description" : "Bad request parameters"
          },
          "500" : {
            "description" : "Error creating new user or sending confirmation e-mail"
          }
        },
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "type" : "object",
                "properties" : {
                  "password" : {
                    "type" : "string",
					"description": "Array of two passwords"
                  },
                  "email" : {
                    "type" : "string"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/users/{idUser}" : {
      "get" : {
        "tags" : [ "Users" ],
        "summary" : "Find user by ID",
        "description" : "Returns a single user matching the ID",
        "operationId" : "getUserByID",
        "parameters" : [ {
          "name" : "idUser",
          "in" : "path",
          "description" : "user's ID",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "User successfully retrieved",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/User"
                }
              }
            }
          },
          "400" : {
            "description" : "Invalid ID supplied"
          },
          "404" : {
            "description" : "User with specified ID not found"
          },
          "500" : {
            "description" : "Error retrieving user"
          }
        }
      },
      "delete" : {
        "tags" : [ "Users" ],
        "summary" : "Delete a user from the database",
        "operationId" : "userDelete",
        "parameters" : [ {
          "name" : "idUser",
          "in" : "path",
          "description" : "User id",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "204" : {
            "description" : "User successfully deleted"
          },
          "400" : {
            "description" : "Invalid ID supplied"
          },
          "404" : {
            "description" : "User with specified ID not founds"
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
      "User" : {
        "type" : "object",
        "properties" : {
          "_id" : {
            "type" : "string",
            "description" : "user's e-mail serves as the id"
          },
          "todoLists" : {
            "type" : "object",
            "description" : "list of user's todo lists"
          },
          "calendars" : {
            "type" : "object",
            "description" : "list of user's calendars"
          },
          "timetables" : {
            "type" : "object",
            "description" : "list of user's timetables"
          },
          "status" : {
            "type" : "integer",
            "description" : "status of account"
          },
          "hashValue" : {
            "type" : "string",
            "description" : "password hash value"
          },
          "randomValue" : {
            "type" : "string",
            "description" : "random value used in hashing"
          },
          "validationCode" : {
            "type" : "string",
            "description" : "value used to generate the address for account validation through e-mail"
          },
          "admin" : {
            "type" : "boolean",
            "description" : "true if administrator else false"
          }
        }
      },
...
    }
  }
}